pyproject.toml
[project]
name = "travel-marketing-agents"
version = "0.1.0"
description = "旅行マーケティング AI マルチエージェントパイプライン"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agent-framework>=1.0.0rc5",
    "azure-ai-contentsafety>=1.0.0",
    "azure-ai-projects>=2.0.1",
    "azure-identity>=1.26.0b2",
    "fastapi>=0.135.2",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.11.0",
]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
